Reworking the placement of metatype groups (#2094)
### Changes

Create a separate module to store metatype groups in one location.

### Reason for changes

Ref: 113580

### Related tickets

Ref: 113580

### Tests

N/A
