+++ nncf/onnx/quantization/default_quantization.py
@@ -11,59 +11,14 @@
 
 from nncf.common.quantization.quantizer_propagation.structs import QuantizationTrait
 from nncf.onnx.graph.metatypes import onnx_metatypes
+from nncf.onnx.graph.metatypes.groups import INPUTS_QUANTIZABLE_OPERATIONS
+from nncf.onnx.graph.metatypes.groups import QUANTIZE_AGNOSTIC_OPERATIONS
 
 # If a metatype is not in this list, then it is considered to be QuantizationTrait.NON_QUANTIZABLE.
 
 DEFAULT_ONNX_QUANT_TRAIT_TO_OP_DICT = {
-    QuantizationTrait.INPUTS_QUANTIZABLE: [
-        onnx_metatypes.ONNXConvolutionMetatype,
-        onnx_metatypes.ONNXDepthwiseConvolutionMetatype,
-        onnx_metatypes.ONNXConvolutionTransposeMetatype,
-        *onnx_metatypes.MATMUL_METATYPES,
-        onnx_metatypes.ONNXAveragePoolMetatype,
-        onnx_metatypes.ONNXGlobalAveragePoolMetatype,
-        onnx_metatypes.ONNXAddLayerMetatype,
-        onnx_metatypes.ONNXSubMetatype,
-        onnx_metatypes.ONNXMulLayerMetatype,
-        onnx_metatypes.ONNXBatchNormMetatype,
-        onnx_metatypes.ONNXHardSigmoidMetatype,
-        onnx_metatypes.ONNXResizeMetatype,
-        onnx_metatypes.ONNXPowMetatype,
-        onnx_metatypes.ONNXReciprocalMetatype,
-        onnx_metatypes.ONNXMaximumMetatype,
-        onnx_metatypes.ONNXMinimumMetatype,
-    ],
-    QuantizationTrait.QUANTIZATION_AGNOSTIC: [
-        onnx_metatypes.ONNXMaxPoolMetatype,
-        onnx_metatypes.ONNXReduceMaxMetatype,
-        onnx_metatypes.ONNXReshapeMetatype,
-        onnx_metatypes.ONNXTransposeMetatype,
-        onnx_metatypes.ONNXSqueezeMetatype,
-        onnx_metatypes.ONNXUnsqueezeMetatype,
-        onnx_metatypes.ONNXSplitMetatype,
-        onnx_metatypes.ONNXTileMetatype,
-        onnx_metatypes.ONNXCenterCropPadMetatype,
-        onnx_metatypes.ONNXSliceMetatype,
-        onnx_metatypes.ONNXPadMetatype,
-        onnx_metatypes.ONNXGatherMetatype,
-        onnx_metatypes.ONNXGatherNDMetatype,
-        onnx_metatypes.ONNXGatherElementsMetatype,
-        onnx_metatypes.ONNXDepthToSpaceMetatype,
-        onnx_metatypes.ONNXSpaceToDepthMetatype,
-        onnx_metatypes.ONNXScatterElementsMetatype,
-        onnx_metatypes.ONNXScatterNDMetatype,
-        onnx_metatypes.ONNXScatterMetatype,
-        onnx_metatypes.ONNXCastLikeMetatype,
-        onnx_metatypes.ONNXDropoutMetatype,
-        onnx_metatypes.ONNXFlattenMetatype,
-        onnx_metatypes.ONNXExpandMetatype,
-        onnx_metatypes.ONNXIdentityMetatype,
-        # ONNXReluMetatype is not considered to be QUANTIZATION_AGNOSTIC, because:
-        # 1. Runtime doesn't provide performance benefits by quantizing the stand-alone RELU's (ticket: 59548)
-        # 2. It's frequently better for the end accuracy to have quantizers set up after the RELU
-        # so that the input distribution to the quantizer is non-negative
-        # and we can therefore have better quantization resolution while preserving the original dynamic range
-    ],
+    QuantizationTrait.INPUTS_QUANTIZABLE: INPUTS_QUANTIZABLE_OPERATIONS,
+    QuantizationTrait.QUANTIZATION_AGNOSTIC: QUANTIZE_AGNOSTIC_OPERATIONS,
     QuantizationTrait.CONCAT: [onnx_metatypes.ONNXConcatMetatype],
     QuantizationTrait.OUTPUT_QUANTIZATION_AS_WEIGHTS: [onnx_metatypes.ONNXEmbeddingMetatype],
 }
