+++ nncf/openvino/graph/metatypes/openvino_metatypes.py
@@ -673,28 +673,6 @@ class OVAbsMetatype(OVOpMetatype):
     op_names = ["Abs"]
 
 
-GENERAL_WEIGHT_LAYER_METATYPES = [
-    OVConvolutionMetatype,
-    OVGroupConvolutionMetatype,
-    OVDepthwiseConvolutionMetatype,
-    OVConvolutionBackpropDataMetatype,
-    OVGroupConvolutionBackpropDataMetatype,
-    OVMatMulMetatype,
-    OVLSTMSequenceMetatype,
-    OVGRUSequenceMetatype,
-    OVEmbeddingMetatype,
-]
-
-METATYPES_WITH_CONST_PORT_ID = GENERAL_WEIGHT_LAYER_METATYPES + [OVAddMetatype]
-
-# Contains the operation metatypes for which bias can be applied.
-OPERATIONS_WITH_BIAS_METATYPES = [
-    OVConvolutionMetatype,
-    # TODO: add all metatypes with bias
-    OVMatMulMetatype,
-]
-
-
 def get_operator_metatypes() -> List[Type[OperatorMetatype]]:
     """
     Returns a list of the operator metatypes.
