+++ nncf/openvino/hardware/fused_patterns.py
@@ -13,11 +13,11 @@ from nncf.common.graph.patterns import GraphPattern
 from nncf.common.graph.patterns import HWFusedPatternNames
 from nncf.common.utils.registry import Registry
 from nncf.openvino.graph.metatypes import openvino_metatypes as om
-from nncf.openvino.hardware.pattern_operations import ARITHMETIC_OPERATIONS
-from nncf.openvino.hardware.pattern_operations import ATOMIC_ACTIVATIONS_OPERATIONS
-from nncf.openvino.hardware.pattern_operations import BATCH_NORMALIZATION_OPERATIONS
-from nncf.openvino.hardware.pattern_operations import ELEMENTWISE_OPERATIONS
-from nncf.openvino.hardware.pattern_operations import LINEAR_OPERATIONS
+from nncf.openvino.graph.metatypes.groups import ARITHMETIC_OPERATIONS
+from nncf.openvino.graph.metatypes.groups import ATOMIC_ACTIVATIONS_OPERATIONS
+from nncf.openvino.graph.metatypes.groups import BATCH_NORMALIZATION_OPERATIONS
+from nncf.openvino.graph.metatypes.groups import ELEMENTWISE_OPERATIONS
+from nncf.openvino.graph.metatypes.groups import LINEAR_OPERATIONS
 
 OPENVINO_HW_FUSED_PATTERNS = Registry("openvino")
 
@@ -110,7 +110,7 @@ def create_normalize() -> GraphPattern:
 @OPENVINO_HW_FUSED_PATTERNS.register(HWFusedPatternNames.LINEAR_WITH_BIAS)
 def create_biased_op() -> GraphPattern:
     pattern = GraphPattern()
-    linear_node = pattern.add_node(**LINEAR_OPERATIONS)
+    linear_node = pattern.add_node(**{GraphPattern.METATYPE_ATTR: LINEAR_OPERATIONS, GraphPattern.LABEL_ATTR: "LINEAR"})
     add_node = pattern.add_node(**{GraphPattern.LABEL_ATTR: "ADD_BIAS", GraphPattern.METATYPE_ATTR: om.OVAddMetatype})
 
     pattern.add_edge(linear_node, add_node)
@@ -156,7 +156,9 @@ def create_se_block() -> GraphPattern:
     reduce_mean_node = pattern.add_node(
         **{GraphPattern.LABEL_ATTR: "REDUCE_MEAN", GraphPattern.METATYPE_ATTR: om.OVReduceMeanMetatype}
     )
-    linear_node_1 = pattern.add_node(**LINEAR_OPERATIONS)
+    linear_node_1 = pattern.add_node(
+        **{GraphPattern.METATYPE_ATTR: LINEAR_OPERATIONS, GraphPattern.LABEL_ATTR: "LINEAR"}
+    )
     add_node_1 = pattern.add_node(**{GraphPattern.LABEL_ATTR: "ADD_BIAS", GraphPattern.METATYPE_ATTR: om.OVAddMetatype})
     activation_node_1 = pattern.add_node(
         **{
@@ -164,7 +166,9 @@ def create_se_block() -> GraphPattern:
             GraphPattern.METATYPE_ATTR: [om.OVReluMetatype, om.OVPReluMetatype, om.OVSwishMetatype],
         }
     )
-    linear_node_2 = pattern.add_node(**LINEAR_OPERATIONS)
+    linear_node_2 = pattern.add_node(
+        **{GraphPattern.METATYPE_ATTR: LINEAR_OPERATIONS, GraphPattern.LABEL_ATTR: "LINEAR"}
+    )
     add_node_2 = pattern.add_node(**{GraphPattern.LABEL_ATTR: "ADD_BIAS", GraphPattern.METATYPE_ATTR: om.OVAddMetatype})
     activation_node_2 = pattern.add_node(
         **{GraphPattern.LABEL_ATTR: "SIGMOID", GraphPattern.METATYPE_ATTR: om.OVSigmoidMetatype}
@@ -206,7 +210,7 @@ def create_softmax_div() -> GraphPattern:
 @OPENVINO_HW_FUSED_PATTERNS.register(HWFusedPatternNames.HSWISH_ACTIVATION)
 def create_hswish() -> GraphPattern:
     pattern = GraphPattern()
-    linear_node = pattern.add_node(**LINEAR_OPERATIONS)
+    linear_node = pattern.add_node(**{GraphPattern.METATYPE_ATTR: LINEAR_OPERATIONS, GraphPattern.LABEL_ATTR: "LINEAR"})
     add_node_1 = pattern.add_node(**{GraphPattern.LABEL_ATTR: "ADD_BIAS", GraphPattern.METATYPE_ATTR: om.OVAddMetatype})
     add_node_2 = pattern.add_node(**{GraphPattern.LABEL_ATTR: "ADD", GraphPattern.METATYPE_ATTR: om.OVAddMetatype})
     clamp_node = pattern.add_node(**{GraphPattern.LABEL_ATTR: "CLAMP", GraphPattern.METATYPE_ATTR: om.OVClampMetatype})
@@ -260,7 +264,7 @@ def create_hswish_pattern_2() -> GraphPattern:
 @OPENVINO_HW_FUSED_PATTERNS.register(HWFusedPatternNames.HSWISH_ACTIVATION_WITHOUT_DENOMINATOR)
 def create_hswish_without_denominator() -> GraphPattern:
     pattern = GraphPattern()
-    linear_node = pattern.add_node(**LINEAR_OPERATIONS)
+    linear_node = pattern.add_node(**{GraphPattern.METATYPE_ATTR: LINEAR_OPERATIONS, GraphPattern.LABEL_ATTR: "LINEAR"})
     add_node_1 = pattern.add_node(**{GraphPattern.LABEL_ATTR: "ADD_BIAS", GraphPattern.METATYPE_ATTR: om.OVAddMetatype})
     add_node_2 = pattern.add_node(**{GraphPattern.LABEL_ATTR: "ADD", GraphPattern.METATYPE_ATTR: om.OVAddMetatype})
     clamp_node = pattern.add_node(**{GraphPattern.LABEL_ATTR: "CLAMP", GraphPattern.METATYPE_ATTR: om.OVClampMetatype})
@@ -279,7 +283,7 @@ def create_hswish_without_denominator() -> GraphPattern:
 @OPENVINO_HW_FUSED_PATTERNS.register(HWFusedPatternNames.SWISH_WITH_HARD_SIGMOID)
 def create_swish_with_hardsigmoid() -> GraphPattern:
     pattern = GraphPattern()
-    linear_node = pattern.add_node(**LINEAR_OPERATIONS)
+    linear_node = pattern.add_node(**{GraphPattern.METATYPE_ATTR: LINEAR_OPERATIONS, GraphPattern.LABEL_ATTR: "LINEAR"})
     add_node = pattern.add_node(**{GraphPattern.LABEL_ATTR: "ADD_BIAS", GraphPattern.METATYPE_ATTR: om.OVAddMetatype})
     hard_sigmoid_node = pattern.add_node(
         **{GraphPattern.LABEL_ATTR: "HARDSIGMOID", GraphPattern.METATYPE_ATTR: om.OVHardSigmoidMetatype}
@@ -715,31 +719,35 @@ def create_linear_biased_activation_elementwise() -> GraphPattern:
 
 def elementwise_operations() -> GraphPattern:
     pattern = GraphPattern()
-    pattern.add_node(**ELEMENTWISE_OPERATIONS)
+    pattern.add_node(**{GraphPattern.METATYPE_ATTR: ELEMENTWISE_OPERATIONS, GraphPattern.LABEL_ATTR: "ELEMENTWISE"})
     return pattern
 
 
 def linear_operations() -> GraphPattern:
     pattern = GraphPattern()
-    pattern.add_node(**LINEAR_OPERATIONS)
+    pattern.add_node(**{GraphPattern.METATYPE_ATTR: LINEAR_OPERATIONS, GraphPattern.LABEL_ATTR: "LINEAR"})
     return pattern
 
 
 def batch_normalization_operations() -> GraphPattern:
     pattern = GraphPattern()
-    pattern.add_node(**BATCH_NORMALIZATION_OPERATIONS)
+    pattern.add_node(
+        **{GraphPattern.METATYPE_ATTR: BATCH_NORMALIZATION_OPERATIONS, GraphPattern.LABEL_ATTR: "BATCH_NORMALIZATION"}
+    )
     return pattern
 
 
 def atomic_activations_operations() -> GraphPattern:
     pattern = GraphPattern()
-    pattern.add_node(**ATOMIC_ACTIVATIONS_OPERATIONS)
+    pattern.add_node(
+        **{GraphPattern.METATYPE_ATTR: ATOMIC_ACTIVATIONS_OPERATIONS, GraphPattern.LABEL_ATTR: "ATOMIC_ACTIVATIONS"}
+    )
     return pattern
 
 
 def arithmetic_operations() -> GraphPattern:
     pattern = GraphPattern()
-    pattern.add_node(**ARITHMETIC_OPERATIONS)
+    pattern.add_node(**{GraphPattern.METATYPE_ATTR: ARITHMETIC_OPERATIONS, GraphPattern.LABEL_ATTR: "ARITHMETIC"})
     return pattern
