+++ tests/onnx/test_layer_attributes.py
@@ -13,7 +13,7 @@ import numpy as np
 import onnx
 import pytest
 
-from nncf.onnx.graph.metatypes.onnx_metatypes import GENERAL_WEIGHT_LAYER_METATYPES
+from nncf.onnx.graph.metatypes.groups import OPERATIONS_WITH_WEIGHTS
 from nncf.onnx.graph.nncf_graph_builder import GraphConverter
 from nncf.onnx.graph.nncf_graph_builder import ONNXLayerAttributes
 from tests.onnx.models import OPSET_VERSION
@@ -111,7 +111,7 @@ def test_layer_attributes(node_creator, ref_layer_attrs):
     onnx_model = get_one_layer_model(op_name, node_creator, input_shape)
     nncf_graph = GraphConverter.create_nncf_graph(onnx_model)
     node = nncf_graph.get_node_by_name(op_name)
-    if node.metatype in GENERAL_WEIGHT_LAYER_METATYPES:
+    if node.metatype in OPERATIONS_WITH_WEIGHTS:
         assert node.layer_attributes.__dict__ == ref_layer_attrs.__dict__
     else:
         assert node.layer_attributes.__dict__ == ONNXLayerAttributes().__dict__
